@@ -19,7 +19,14 @@
   },
   "resolutions": {
     "file-saver": "2.0.5",
-    "rxjs": "7.5.5"
+    "chokidar": "^3",
+    "fsevents": "^2",
+    "mssql": "^8.1.0",
+    "node-addon-api": "^5",
+    "nx": "^14",
+    "rxjs": "7.5.5",
+    "sharp": "^0.30.4",
+    "uuid": "^8"
   },
   "dependencies": {
     "@docusaurus/core": "2.0.0-beta.13",
@@ -28,9 +35,10 @@
     "clsx": "^1.1.1",
     "concurrently": "^7.1.0",
     "env-cmd": "^10.1.0",
+    "global": "^4.4.0",
     "prism-react-renderer": "^1.2.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
     "tslib": "^2.3.0"
   },
   "devDependencies": {
@@ -41,23 +49,22 @@
     "@commitlint/config-conventional": "^16.2.4",
     "@docusaurus/module-type-aliases": "2.0.0-beta.13",
     "@jscutlery/semver": "^2.23.3",
-    "@nrwl/angular": "^14.0.5",
-    "@nrwl/cli": "14.0.5",
-    "@nrwl/cypress": "^14.0.5",
-    "@nrwl/devkit": "^14.0.5",
-    "@nrwl/eslint-plugin-nx": "^14.0.5",
-    "@nrwl/jest": "^14.0.5",
-    "@nrwl/js": "^14.0.5",
-    "@nrwl/linter": "^14.0.5",
-    "@nrwl/nest": "^14.0.5",
-    "@nrwl/node": "^14.0.5",
-    "@nrwl/nx-cloud": "latest",
-    "@nrwl/nx-plugin": "^14.0.5",
-    "@nrwl/react": "^14.0.5",
-    "@nrwl/storybook": "^14.0.5",
-    "@nrwl/tao": "^14.0.5",
-    "@nrwl/web": "^14.0.5",
-    "@nrwl/workspace": "^14.0.5",
+    "@nrwl/angular": "14.1.5",
+    "@nrwl/cli": "14.1.5",
+    "@nrwl/cypress": "14.1.5",
+    "@nrwl/devkit": "14.1.5",
+    "@nrwl/eslint-plugin-nx": "14.1.5",
+    "@nrwl/jest": "14.1.5",
+    "@nrwl/js": "14.1.5",
+    "@nrwl/linter": "14.1.5",
+    "@nrwl/nest": "14.1.5",
+    "@nrwl/node": "14.1.5",
+    "@nrwl/nx-cloud": "14.0.3",
+    "@nrwl/nx-plugin": "14.1.5",
+    "@nrwl/react": "14.1.5",
+    "@nrwl/storybook": "14.1.5",
+    "@nrwl/web": "14.1.5",
+    "@nrwl/workspace": "14.1.5",
     "@nx-clean/plugin-core": "^1.7.0",
     "@nx-dotnet/core": "^1.9.12",
     "@nx-go/nx-go": "^2.5.0",
@@ -71,20 +78,22 @@
     "@typescript-eslint/parser": "~5.18.0",
     "@yarnpkg/plugin-dlx": "^4.0.0-rc.4",
     "all-contributors-cli": "^6.20.0",
-    "dotenv": "^16.0.0",
+    "dotenv": "10.0.0",
     "esbuild": "^0.14.38",
     "eslint": "~8.12.0",
     "eslint-config-prettier": "^8.5.0",
     "husky": "^7.0.4",
     "jest": "27.5.1",
     "jest-extended": "^2.0.0",
+    "nx": "14.1.5",
     "nx-plugin-esbuild": "^0.5.0",
+    "nxpm": "^1.15.0",
     "pm2": "^5.2.0",
-    "prettier": "^2.3.1",
+    "prettier": "2.6.2",
     "ts-jest": "27.1.4",
     "ts-node": "9.1.1",
     "tslib": "^2.0.0",
-    "typescript": "~4.3.5",
+    "typescript": "4.3.5",
     "webpack": "^5.72.0"
   },
   "packageManager": "yarn@3.2.0",
